[Test] Make test suite independent from torchvision (#1252)
* initial commit

* initial commit

* cleanup
