+++ tests/conftest.py
@@ -21,7 +21,7 @@ from typing import List
 
 import pytest
 from tests.helpers import delete_file
-from tests.utils.torch import find_file_with_pattern
+from tests.utils.helpers import find_file_with_pattern
 
 
 logging.basicConfig(level=logging.INFO)
