+++ tests/utils/helpers.py
@@ -1,0 +1,25 @@
+# Copyright (c) 2021 - present / Neuralmagic, Inc. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import os
+import re
+from typing import Optional
+
+
+def find_file_with_pattern(folder_path: str, pattern: str) -> Optional[str]:
+    folder_path = os.path.expanduser(folder_path)
+    for root, dirs, files in os.walk(folder_path):
+        for file in files:
+            if re.match(pattern, file):
+                return os.path.join(root, file)
